🚚 Split `many` module into submodules
One for each combinator. The module size was getting out of control (1.2k+ LoC).
In addition, bring `many_until` inside the `many` module.
