+++ src/combinators/many/many0_discard.rs
@@ -1,0 +1,176 @@
+use super::core::{ManyOutput, many};
+use crate::{InputToken, Parser};
+
+/// Applies `parser` zero or more time, without collecting possible matches.
+///
+/// # Outcome
+///
+/// This parser always succeeds, even if its argument parser doesn't. Unlike [`crate::many0`], this
+/// combinator doesn't return its matches, but just how many times it matched.
+///
+/// # Input consumption
+///
+/// This parser consumes input if:
+/// - At least one occurrence of its argument parser succeeds.
+/// - If its argument parser consumes input upon failure, independent of this combinator's outcome.
+///
+/// # Error handling
+///
+/// This combinator fails with [`crate::Error::NonConsumingLoop`] if the argument parser does not
+/// consume input upon success. This behavior is there to prevent an infinite loop caused by the
+/// input never being consumed.
+///
+/// # Look-ahead and backtracking
+///
+/// This combinator doesn't perform any lookahead. It also never backtracks, given that it never
+/// fails.
+///
+/// # Shortcut
+///
+/// This combinator has a shortcut version: [`Parser::many0_discard`].
+///
+/// # Arguments
+///
+/// - `parser`: The parser to possibly be applied many times.
+///
+/// # Examples
+///
+/// ```
+/// use yapcol::{Input, is, many0_discard};
+///
+/// // Matches multiple elements
+/// let parser = is('1');
+/// let mut input = Input::new_from_chars("112".chars(), None);
+/// assert_eq!(many0_discard(&parser)(&mut input), Ok(2));
+///
+/// // Returns an empty vector when no matches are found (never fails)
+/// let mut input = Input::new_from_chars("23".chars(), None);
+/// assert_eq!(many0_discard(&parser)(&mut input), Ok(0));
+///
+/// // Returns an empty vector on empty input (never fails)
+/// let mut input = Input::new_from_chars("".chars(), None);
+/// assert_eq!(many0_discard(&parser)(&mut input), Ok(0));
+/// ```
+pub fn many0_discard<P, IT, O>(parser: &P) -> impl Parser<IT, usize>
+where
+	P: Parser<IT, O>,
+	IT: InputToken,
+{
+	move |input| match many(parser, 0, None, false)(input) {
+		Ok(ManyOutput::Matches(_)) => panic!("[many0_discard] Expected Count, but got Matches."),
+		Ok(ManyOutput::Count(count)) => Ok(count),
+		Err(e) => Err(e),
+	}
+}
+
+#[cfg(test)]
+mod tests {
+	use crate::input::Position;
+	use crate::*;
+
+	#[test]
+	fn empty() {
+		let parser = is('h');
+		let mut input = Input::new_from_chars("".chars(), None);
+		let parser_many0 = many0_discard(&parser);
+		let output = parser_many0(&mut input).unwrap();
+		assert_eq!(output, 0);
+	}
+
+	#[test]
+	fn empty_shortcut() {
+		let parser = is('h').many0_discard();
+		let mut input = Input::new_from_chars("".chars(), None);
+		let output = parser(&mut input).unwrap();
+		assert_eq!(output, 0);
+	}
+
+	#[test]
+	fn no_match_not_empty() {
+		let token_count = 100;
+		let parser = is('h');
+		let tokens = std::iter::repeat_n('j', token_count).collect::<Vec<_>>();
+		let mut input = Input::new_from_chars(tokens, None);
+		let parser_many0 = many0_discard(&parser);
+		let output = parser_many0(&mut input).unwrap();
+		assert_eq!(output, 0);
+		assert_eq!(input.consumed_count(), 0);
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+	}
+
+	#[test]
+	fn no_match_not_empty_shortcut() {
+		let token_count = 100;
+		let parser = is('h').many0_discard();
+		let tokens = std::iter::repeat_n('j', token_count).collect::<Vec<_>>();
+		let mut input = Input::new_from_chars(tokens, None);
+		let output = parser(&mut input).unwrap();
+		assert_eq!(output, 0);
+		assert_eq!(input.consumed_count(), 0);
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+	}
+
+	#[test]
+	fn match_not_empty() {
+		let token_count = 100;
+		let parser = is('h');
+		let tokens = std::iter::repeat_n('h', token_count).collect::<Vec<_>>();
+		let mut input = Input::new_from_chars(tokens, None);
+		let parser_many0 = many0_discard(&parser);
+		let output = parser_many0(&mut input).unwrap();
+		assert_eq!(output, token_count);
+		assert_eq!(input.consumed_count(), token_count);
+		assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
+	}
+
+	#[test]
+	fn match_not_empty_shortcut() {
+		let token_count = 100;
+		let parser = is('h').many0_discard();
+		let tokens = std::iter::repeat_n('h', token_count).collect::<Vec<_>>();
+		let mut input = Input::new_from_chars(tokens, None);
+		let output = parser(&mut input).unwrap();
+		assert_eq!(output, token_count);
+		assert_eq!(input.consumed_count(), token_count);
+		assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
+	}
+
+	#[test]
+	fn partial_match_then_stop() {
+		let parser = is('#');
+		let mut input = Input::new_from_chars("#####Hello".chars(), None);
+		let parser_many0 = many0_discard(&parser);
+		let output = parser_many0(&mut input).unwrap();
+		assert_eq!(output, 5);
+		assert_eq!(input.consumed_count(), 5);
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+		assert_eq!(any()(&mut input), Ok('H'));
+	}
+
+	#[test]
+	fn non_consuming_parser_does_not_loop() {
+		let parser = success(1); // Non-consuming parser.
+		let mut input = Input::new_from_chars("hello".chars(), None);
+		let parser = parser.many0_discard();
+		let output = parser(&mut input);
+		let position = Position::new(1, 1);
+		assert_eq!(output, Err(Error::NonConsumingLoop(None, position)));
+	}
+
+	#[test]
+	fn match_consuming_upon_failure() {
+		// Parser that consumes input upon failure:
+		let parser = |input: &mut StringInput| {
+			let o1 = is('#')(input)?;
+			let o2 = is('a')(input)?;
+			Ok((o1, o2))
+		};
+		let mut input = Input::new_from_chars("#a#e".chars(), None);
+		let many_parser = parser.many0_discard();
+		let output = many_parser(&mut input).unwrap();
+		assert_eq!(output, 1);
+		assert_eq!(input.consumed_count(), 3); // The second attempt failed while consuming.
+		assert_eq!(any()(&mut input), Ok('e'));
+		assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
+	}
+}
