+++ src/combinators/many/many1_up_to.rs
@@ -1,0 +1,191 @@
+use super::core::{ManyOutput, many};
+use crate::{InputToken, Parser};
+
+/// Applies `parser` between 1 and a given number of times, ensuring that no more matches occur.
+///
+/// # Outcome
+///
+/// This combinator succeeds if the argument parser succeeds between 1 and up to (and including)
+/// `max_count` times. In that case, it returns a vector of matches of its argument parser.
+///
+/// It fails if the argument parser:
+/// - Never succeeds.
+/// - Matches more than `max_count` times.
+///
+/// # Input consumption
+///
+/// This parser consumes input if:
+/// - At least one occurrence of its argument parser succeeds.
+/// - If its argument parser consumes input upon failure, independent of this combinator's outcome.
+///
+/// # Error handling
+///
+/// This combinator fails with [`crate::Error::NonConsumingLoop`] if the argument parser does not
+/// consume input upon success. This behavior is there to prevent an infinite loop caused by the
+/// input never being consumed.
+///
+/// # Look-ahead and backtracking
+///
+/// This combinator doesn't perform any lookahead and won't backtrack upon failure.
+///
+/// # Shortcut
+///
+/// This combinator has a shortcut version: [`Parser::many1_up_to`].
+///
+/// # Arguments
+///
+/// - `parser`: The parser to be applied multiple times.
+/// - `max_count`: The (inclusive) maximum number of times that the argument parser should succeed.
+///   Must be greater than 0, otherwise this function panics.
+///
+/// # Panics
+///
+/// This function panics if `max_count` is equal to 0. Check [`crate::many0_up_to`] if you would
+/// like to cover this case.
+///
+/// # Examples
+///
+/// ```
+/// use yapcol::{Input, is, many1_up_to};
+///
+/// // Succeeds if the parser matches exactly `max_count` times.
+/// let parser = is('1');
+/// let mut input = Input::new_from_chars("112".chars(), None);
+/// let max_count = 2;
+/// assert_eq!(
+/// 	many1_up_to(&parser, max_count)(&mut input),
+/// 	Ok("11".chars().collect())
+/// );
+///
+/// // Succeeds if the parser matches less than `max_count` times.
+/// let parser = is('1');
+/// let mut input = Input::new_from_chars("112".chars(), None);
+/// let max_count = 5;
+/// assert_eq!(
+/// 	many1_up_to(&parser, max_count)(&mut input),
+/// 	Ok("11".chars().collect())
+/// );
+///
+/// // Fails if the parser matches more than `max_count` times.
+/// let parser = is('1');
+/// let mut input = Input::new_from_chars("1112".chars(), None);
+/// let max_count = 2;
+/// assert!(many1_up_to(&parser, max_count)(&mut input).is_err());
+/// ```
+pub fn many1_up_to<P, IT, O>(parser: &P, max_count: usize) -> impl Parser<IT, Vec<O>>
+where
+	P: Parser<IT, O>,
+	IT: InputToken,
+{
+	if max_count == 0 {
+		panic!("max_count must be greater than 0");
+	}
+	move |input| match many(parser, 1, Some(max_count), true)(input) {
+		Ok(ManyOutput::Matches(matches)) => Ok(matches),
+		Ok(ManyOutput::Count(_)) => panic!("[many1_up_to] Expected Matches, but got Count."),
+		Err(e) => Err(e),
+	}
+}
+
+#[cfg(test)]
+mod tests {
+	use super::*;
+	use crate::combinators::many::test_utils::assert_unexpected_error;
+	use crate::input::Position;
+	use crate::*;
+
+	#[test]
+	fn empty() {
+		let parser = is('h');
+		let mut input = Input::new_from_chars("".chars(), None);
+		let parser_up_to = many1_up_to(&parser, 1);
+		assert_eq!(
+			parser_up_to(&mut input),
+			Err(Error::EndOfInput(Some(Box::new('h'))))
+		);
+	}
+
+	#[test]
+	fn empty_shortcut() {
+		let parser = is('h').many1_up_to(1);
+		let mut input = Input::new_from_chars("".chars(), None);
+		assert_eq!(
+			parser(&mut input),
+			Err(Error::EndOfInput(Some(Box::new('h'))))
+		);
+	}
+
+	#[test]
+	fn no_match() {
+		let parser = is('h');
+		let mut input = Input::new_from_chars("jklmno".chars(), None);
+		let parser_up_to = many1_up_to(&parser, 1);
+		let output = parser_up_to(&mut input);
+		let position = Position::new(1, 1);
+		assert_unexpected_error(output, position, "h", "j");
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+	}
+
+	#[test]
+	fn no_match_shortcut() {
+		let parser = is('h').many1_up_to(1);
+		let mut input = Input::new_from_chars("jklmno".chars(), None);
+		let output = parser(&mut input);
+		let position = Position::new(1, 1);
+		assert_unexpected_error(output, position, "h", "j");
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+	}
+
+	#[test]
+	fn one_match() {
+		let parser = is('h');
+		let mut input = Input::new_from_chars("hello".chars(), None);
+		let parser_up_to = many1_up_to(&parser, 1);
+		let output = parser_up_to(&mut input).unwrap();
+		assert_eq!(output, vec!['h']);
+		assert_eq!(input.consumed_count(), 1);
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+		assert_eq!(any()(&mut input), Ok('e'));
+	}
+
+	#[test]
+	fn less_than_max_count() {
+		let parser = is('h');
+		let mut input = Input::new_from_chars("hhello".chars(), None);
+		let parser_up_to = many1_up_to(&parser, 3);
+		let output = parser_up_to(&mut input).unwrap();
+		assert_eq!(output, vec!['h', 'h']);
+		assert_eq!(input.consumed_count(), 2);
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+		assert_eq!(any()(&mut input), Ok('e'));
+	}
+
+	#[test]
+	fn equal_to_max_count() {
+		let parser = is('h');
+		let mut input = Input::new_from_chars("hhhello".chars(), None);
+		let parser_up_to = many1_up_to(&parser, 3);
+		let output = parser_up_to(&mut input).unwrap();
+		assert_eq!(output, vec!['h', 'h', 'h']);
+		assert_eq!(input.consumed_count(), 3);
+		assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
+		assert_eq!(any()(&mut input), Ok('e'));
+	}
+
+	#[test]
+	fn more_than_max_count() {
+		let parser = is('h');
+		let mut input = Input::new_from_chars("hhhhello".chars(), None);
+		let parser_up_to = many1_up_to(&parser, 3);
+		let output = parser_up_to(&mut input);
+		let position = Position::new(1, 4);
+		assert_unexpected_error(output, position, "3", "4");
+	}
+
+	#[test]
+	#[should_panic]
+	fn zero_panics() {
+		let parser = is::<CharToken>('h');
+		let _ = many1_up_to(&parser, 0);
+	}
+}
