+++ src/combinators/many/mod.rs
@@ -1,0 +1,18 @@
+mod core;
+mod many0;
+mod many0_discard;
+mod many0_up_to;
+mod many1;
+mod many1_discard;
+mod many1_up_to;
+mod many_until;
+#[cfg(test)]
+mod test_utils;
+
+pub use many_until::many_until;
+pub use many0::many0;
+pub use many0_discard::many0_discard;
+pub use many0_up_to::many0_up_to;
+pub use many1::many1;
+pub use many1_discard::many1_discard;
+pub use many1_up_to::many1_up_to;
