+++ src/combinators/mod.rs
@@ -11,7 +11,6 @@ mod end_of_input;
 mod is;
 mod look_ahead;
 mod many;
-mod many_until;
 mod maybe;
 mod not_followed_by;
 mod satisfy;
@@ -28,8 +27,7 @@ pub use either::either;
 pub use end_of_input::end_of_input;
 pub use is::is;
 pub use look_ahead::look_ahead;
-pub use many::{many0, many0_discard, many0_up_to, many1, many1_discard, many1_up_to};
-pub use many_until::many_until;
+pub use many::{many_until, many0, many0_discard, many0_up_to, many1, many1_discard, many1_up_to};
 pub use maybe::maybe;
 pub use not_followed_by::not_followed_by;
 pub use satisfy::satisfy;
